#53 게시판 조회수

+++ src/main/java/com/dongyoung/noAlone/board/model/FindResponseBoardListModel.java
@@ -6,7 +6,7 @@ public record FindResponseBoardListModel(
         Long boardId,
         String title,
         String content,
-        Long views,
+        Integer views,
         LocalDate inputDt
 ) {
 }
